Fix #72, spacing around symbols
Root cause of the problem was in App.funcs.fixIt(), where match replacements are counted. In the case of the symbol_then_space rule, the expression matched properly on a large sample of text. However, in App.funcs.fixIt() the separate matches are replaced individually so that they can be counted, leaving out matches that didn't result in changes. The symbol_then_space rule does not recognize its previous matches for isolated replace operations.